+++ packages/react-reconciler/src/ReactFiberClassUpdateQueue.new.js
@@ -109,7 +109,7 @@ import {
   markSkippedUpdateLanes,
   isUnsafeClassRenderPhaseUpdate,
 } from './ReactFiberWorkLoop.new';
-import {pushInterleavedQueue} from './ReactFiberInterleavedUpdates.new';
+import {pushConcurrentUpdateQueue} from './ReactFiberConcurrentUpdates.new';
 import {setIsStrictModeForDevtools} from './ReactFiberDevToolsHook.new';
 
 import assign from 'shared/assign';
@@ -242,7 +242,7 @@ export function enqueueUpdate<State>(
       update.next = update;
       // At the end of the current render, this queue's interleaved updates will
       // be transferred to the pending queue.
-      pushInterleavedQueue(sharedQueue);
+      pushConcurrentUpdateQueue(sharedQueue);
     } else {
       update.next = interleaved.next;
       interleaved.next = update;
