+++ packages/react-reconciler/src/ReactFiberConcurrentUpdates.old.js
@@ -7,36 +7,36 @@
  * @flow
  */
 
-import type {UpdateQueue as HookQueue} from './ReactFiberHooks.new';
-import type {SharedQueue as ClassQueue} from './ReactUpdateQueue.new';
+import type {UpdateQueue as HookQueue} from './ReactFiberHooks.old';
+import type {SharedQueue as ClassQueue} from './ReactFiberClassUpdateQueue.old';
 
 // An array of all update queues that received updates during the current
 // render. When this render exits, either because it finishes or because it is
 // interrupted, the interleaved updates will be transferred onto the main part
 // of the queue.
-let interleavedQueues: Array<
+let concurrentQueues: Array<
   HookQueue<any, any> | ClassQueue<any>,
 > | null = null;
 
-export function pushInterleavedQueue(
+export function pushConcurrentUpdateQueue(
   queue: HookQueue<any, any> | ClassQueue<any>,
 ) {
-  if (interleavedQueues === null) {
-    interleavedQueues = [queue];
+  if (concurrentQueues === null) {
+    concurrentQueues = [queue];
   } else {
-    interleavedQueues.push(queue);
+    concurrentQueues.push(queue);
   }
 }
 
-export function enqueueInterleavedUpdates() {
+export function finishQueueingConcurrentUpdates() {
   // Transfer the interleaved updates onto the main queue. Each queue has a
   // `pending` field and an `interleaved` field. When they are not null, they
   // point to the last node in a circular linked list. We need to append the
   // interleaved list to the end of the pending list by joining them into a
   // single, circular list.
-  if (interleavedQueues !== null) {
-    for (let i = 0; i < interleavedQueues.length; i++) {
-      const queue = interleavedQueues[i];
+  if (concurrentQueues !== null) {
+    for (let i = 0; i < concurrentQueues.length; i++) {
+      const queue = concurrentQueues[i];
       const lastInterleavedUpdate = queue.interleaved;
       if (lastInterleavedUpdate !== null) {
         queue.interleaved = null;
@@ -50,6 +50,6 @@ export function enqueueInterleavedUpdates() {
         queue.pending = (lastInterleavedUpdate: any);
       }
     }
-    interleavedQueues = null;
+    concurrentQueues = null;
   }
 }
